requirements.txt
torch==2.1.0
gradio==3.25.0
Pillow==9.3.0
torchvision==0.16.0

